fix(security): resolve 16 CodeQL scanning alerts and vulnerabilities
Perform a major security cleanup across the repository to resolve all 16 CodeQL alerts:
- Fix Path Traversal vulnerabilities in Dev Server (cmd/servir.go) by validating absolute paths.
- Prevent integer conversion and overflow vulnerabilities in multiple stdlib files and hrp by adding boundary validations and using safe comparison types.
- Fix JS string replacement escaping in VS Code extension (vscode-harpia/extension.js).
- Add explicit contents: read permissions to GitHub Actions workflows (ci.yml, publicar-extensao.yml).

+++ hrp/bytes.go
@@ -81,7 +81,7 @@ func (b *Bytes) M__maior_ou_igual__(outro Objeto) (Objeto, error) {
 		return nil, err
 	}
 
-	return NewBooleano(len(b.Itens) >= int(outroT.(Inteiro)))
+	return NewBooleano(int64(len(b.Itens)) >= int64(outroT.(Inteiro)))
 }
 
 // M__maior_que__ compara os comprimentos físicos dos bytes.
@@ -91,7 +91,7 @@ func (b *Bytes) M__maior_que__(outro Objeto) (Objeto, error) {
 		return nil, err
 	}
 
-	return NewBooleano(len(b.Itens) > int(outroT.(Inteiro)))
+	return NewBooleano(int64(len(b.Itens)) > int64(outroT.(Inteiro)))
 }
 
 // M__menor_ou_igual__ compara os comprimentos físicos dos bytes.
@@ -101,7 +101,7 @@ func (b *Bytes) M__menor_ou_igual__(outro Objeto) (Objeto, error) {
 		return nil, err
 	}
 
-	return NewBooleano(len(b.Itens) <= int(outroT.(Inteiro)))
+	return NewBooleano(int64(len(b.Itens)) <= int64(outroT.(Inteiro)))
 }
 
 // M__menor_que__ compara os comprimentos físicos dos bytes.
@@ -111,7 +111,7 @@ func (b *Bytes) M__menor_que__(outro Objeto) (Objeto, error) {
 		return nil, err
 	}
 
-	return NewBooleano(len(b.Itens) < int(outroT.(Inteiro)))
+	return NewBooleano(int64(len(b.Itens)) < int64(outroT.(Inteiro)))
 }
 
 // M__tamanho__ retorna o comprimento físico total de bytes armazenados no array local.
